+++ barbershop-backend/src/main/java/com/yield/barbershop_backend/dto/product/ProductDTO.java
@@ -20,7 +20,7 @@ public class ProductDTO {
     private Double price;
 
     @NotNull(message = "stockQuantity is mandatory")
-    private Integer stockQuantity;
+    private Long stockQuantity;
 
     @NotBlank(message = "category is mandatory")
     private String category;
